+++ Src/DeUrgenta.Api/Startup.cs
@@ -21,9 +21,9 @@ using Microsoft.Extensions.DependencyInjection;
 using DeUrgenta.Certifications.Api;
 using FluentValidation.AspNetCore;
 using DeUrgenta.Admin.Api;
+using DeUrgenta.Emailing.Service;
 using DeUrgenta.Events.Api;
 using DeUrgenta.Events.Api.Controller;
-using DeUrgenta.Services.Emailing;
 
 namespace DeUrgenta.Api
 {
