+++ Src/DeUrgenta.Api/appsettings.json
@@ -15,22 +15,26 @@
     "Secret": "llvudfvkwvepwkdnsnwmuulyvtrawppf",
     "TokenExpirationInSeconds": 3600
   },
-  "EmailingSystem": "SendGrid",
-  "AdminFromName": "Echipa 'de urgenta'",
-  "AdminFromEmail": "admin@deurgenta.ro",
   "ConfirmationUrl": "https://api.deurgenta.hostmysite.ro/confirmEmail.html",
   "ResetUrl": "https://api.deurgenta.hostmysite.ro/resetPassword.html",
-  "Smtp": {
-    "Host": "",
-    "Port": 0,
-    "User": "",
-    "Password": ""
+  "Email": {
+    "EmailingSystem": "SendGrid",
+    "AdminFromName": "Echipa 'de urgenta'",
+    "AdminFromEmail": "admin@deurgenta.ro",
+    "Options": {
+      "TemplateFolder": "EmailTemplates"
+    },
+    "Smtp": {
+      "Host": "",
+      "Port": 0,
+      "User": "",
+      "Password": ""
+    },
+    "SendGrid": {
+      "ApiKey": "",
+      "ClickTracking": false
+    }
   },
-  "SendGrid": {
-    "ApiKey": "",
-    "ClickTracking": false
-  },
-  "TemplateFolder": "EmailTemplates",
   "StorageService": "Local",
   "LocalConfigOptions": {
     "Path": "LocalCertificationStore",
