+++ Src/DeUrgenta.Services.Emailing/Builders/IEmailBuilderService.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
-using DeUrgenta.Services.Emailing.Models;
+using DeUrgenta.Emailing.Service.Models;
 
-namespace DeUrgenta.Services.Emailing.Services
+namespace DeUrgenta.Emailing.Service.Builders
 {
     public interface IEmailBuilderService
     {
