+++ Src/DeUrgenta.Services.Emailing/Config/EmailOptions.cs
@@ -1,0 +1,8 @@
+﻿namespace DeUrgenta.Emailing.Service.Config
+{
+    public class EmailOptions
+    {
+        public string TemplateFolder { get; set; }
+       
+    }
+}
