+++ Src/DeUrgenta.Services.Emailing/DeUrgenta.Emailing.Service.csproj
