+++ Src/DeUrgenta.Services.Emailing/Senders/BaseEmailSender.cs
@@ -1,8 +1,9 @@
 using System.Threading;
 using System.Threading.Tasks;
-using DeUrgenta.Services.Emailing.Models;
+using DeUrgenta.Emailing.Service.Builders;
+using DeUrgenta.Emailing.Service.Models;
 
-namespace DeUrgenta.Services.Emailing.Services
+namespace DeUrgenta.Emailing.Service.Senders
 {
     public abstract class BaseEmailSender : IEmailSender
     {
