--- Src/DeUrgenta.Services.Emailing/Services/ITemplateFileSelector.cs
@@ -1,9 +1,0 @@
-﻿using DeUrgenta.Services.Emailing.Models;
-
-namespace DeUrgenta.Services.Emailing.Services
-{
-    public interface ITemplateFileSelector
-    {
-        string GetTemplatePath(EmailTemplate template);
-    }
-}
